Add data generator

+++ demomysql/src/main/java/xyz/hemoreg/demomysql/usermanagment/dto/ClientDTO.java
@@ -6,12 +6,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ClientDTO {
-    private int id;
+    private long id;
     private String name;
     private String email;
     private int age;
 
-    public ClientDTO(int id, String name, String email, int age) {
+    public ClientDTO(long id, String name, String email, int age) {
         this.id = id;
         this.name = name;
         this.email = email;
