+++ demopostgresql/src/main/java/xyz/hemoreg/demopostgresql/usermanagment/dto/ProductDTO.java
@@ -6,11 +6,11 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ProductDTO {
-    private int id;
+    private long id;
     private String name;
     private double price;
 
-    public ProductDTO(int id, String name, double price) {
+    public ProductDTO(long id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
